+++ source/features/search/components/EpochsSearchResult.tsx
@@ -1,0 +1,91 @@
+import { Observer } from 'mobx-react-lite';
+import { useRouter } from 'next/router';
+import React from 'react';
+import Container from '../../../widgets/container/Container';
+import BlockList from '../../blocks/ui/BlockList';
+import EpochSummary from '../../epochs/ui/EpochSummary';
+import StakeDistribution from '../../epochs/ui/StakeDistribution';
+import { useSearchFeature } from '../index';
+import styles from './EpochsSearchResult.scss';
+
+const stakeDistribution = [
+  {
+    slotsElectedPercentage: 1,
+    stakePool: 'Help the USA Cats',
+    stakePoolName: 'CATS',
+  },
+  {
+    slotsElectedPercentage: 0.9,
+    stakePool: 'Cardano Foundation 1',
+    stakePoolName: 'CF1',
+  },
+  {
+    slotsElectedPercentage: 0.78,
+    stakePool: 'Blush Pool 1',
+    stakePoolName: 'BLS1',
+  },
+  {
+    slotsElectedPercentage: 0.5,
+    stakePool: 'Blush Pool 2',
+    stakePoolName: 'BLS1',
+  },
+  {
+    slotsElectedPercentage: 0.17,
+    stakePool: 'Micro Mining',
+    stakePoolName: 'MNG',
+  },
+  {
+    slotsElectedPercentage: 0.08,
+    stakePool: 'Saint-Petersburg Acade',
+    stakePoolName: 'SPBA',
+  },
+];
+
+export const EpochsSearchResult = () => {
+  const { actions, store } = useSearchFeature();
+  const router = useRouter();
+  const { query } = router;
+  if (query && query.number) {
+    const num = parseInt(query.number as string, 10);
+    actions.searchForEpochByNumber.trigger({ number: num });
+  }
+  return (
+    <Observer>
+      {() => {
+        const { epochSearchResult } = store;
+        if (store.isSearching) {
+          return <div>Is searching</div>;
+        } else if (epochSearchResult) {
+          return (
+            <Container>
+              <div className={styles.epochSummary}>
+                <EpochSummary title="Epoch Summary" epoch={epochSearchResult} />
+              </div>
+              <div className={styles.blockList}>
+                <BlockList
+                  title="Blocks"
+                  items={epochSearchResult.blocks.slice(0, 10).map(b => ({
+                    ...b,
+                    createdAt: Date.now(),
+                    createdBy: 'basfss',
+                    epoch: epochSearchResult.number,
+                    output: 0,
+                    transactions: 0,
+                  }))}
+                />
+              </div>
+              <div className={styles.stakeDistribution}>
+                <StakeDistribution
+                  title="Stake Distribution"
+                  items={stakeDistribution}
+                />
+              </div>
+            </Container>
+          );
+        } else {
+          return <div>Nothing found</div>;
+        }
+      }}
+    </Observer>
+  );
+};