+++ source/lib/types.ts
@@ -1,0 +1,3 @@
+export function isNotNull<T>(b: T | null): b is T {
+  return b != null;
+}