+++ source/pages/block.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
 import NoSSR from 'react-no-ssr';
+import { BrandType } from '../constants';
 import { environment } from '../environment';
+import { BlocksSearchResult } from '../features/search/components/BlocksSearchResult';
+import { SearchFeatureProvider } from '../features/search/components/SearchFeatureProvider';
 import { Footer, Header, Layout } from '../widgets/layout';
 
 let BlockPage = () => <NoSSR />;
@@ -8,7 +11,10 @@ if (environment.IS_CLIENT) {
   BlockPage = () => (
     <NoSSR>
       <Layout>
-        <Header />
+        <Header brandType={BrandType.SHRINKED} />
+        <SearchFeatureProvider>
+          <BlocksSearchResult />
+        </SearchFeatureProvider>
         <Footer />
       </Layout>
     </NoSSR>