+++ source/pages/index.tsx
@@ -1,4 +1,3 @@
-import { noop } from 'lodash';
 import Head from 'next/head';
 import React from 'react';
 import NoSSR from 'react-no-ssr';
@@ -8,6 +7,8 @@ import { BlocksFeatureProvider } from '../features/blocks/ui/BlocksFeatureProvid
 import { LatestBlocks } from '../features/blocks/ui/LatestBlocks';
 import { EpochsFeatureProvider } from '../features/epochs/ui/EpochsFeatureProvider';
 import { LatestEpochs } from '../features/epochs/ui/LatestEpochs';
+import { SearchFeatureProvider } from '../features/search/components/SearchFeatureProvider';
+import { SearchBar } from '../features/search/containers/SearchBar';
 import { isMobileScreen } from '../helpers';
 import { Footer, Header, Layout } from '../widgets/layout';
 import styles from './index.scss';
@@ -37,11 +38,12 @@ if (environment.IS_CLIENT) {
             <SideBackgroundImage className={styles.sideBackgroundImage} />
           </div>
         )}
-        <Header
-          withSearch
-          brandType={BrandType.ENLARGED}
-          searchProps={{ onSearch: noop }}
-        />
+        <Header brandType={BrandType.ENLARGED} />
+        <div>
+          <SearchFeatureProvider>
+            <SearchBar brandType={BrandType.ENLARGED} />
+          </SearchFeatureProvider>
+        </div>
         <div className={styles.epochList}>
           <EpochsFeatureProvider>
             <LatestEpochs />