+++ source/pages/transaction.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import NoSSR from 'react-no-ssr';
+import { BrandType } from '../constants';
 import { environment } from '../environment';
 import { Footer, Header, Layout } from '../widgets/layout';
 
@@ -8,7 +9,7 @@ if (environment.IS_CLIENT) {
   TransactionPage = () => (
     <NoSSR>
       <Layout>
-        <Header />
+        <Header brandType={BrandType.SHRINKED} />
         <Footer />
       </Layout>
     </NoSSR>