+++ stories/footer.stories.tsx
@@ -1,9 +1,8 @@
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import { BrandType } from '../source/constants';
-import { Header } from '../source/widgets/layout';
+import { Footer } from '../source/widgets/layout';
 import { PaddingDecorator } from './support/PaddingDecorator';
 
-storiesOf('Header', module)
+storiesOf('Footer', module)
   .addDecorator(story => <PaddingDecorator>{story()}</PaddingDecorator>)
-  .add('Header', () => <Header brandType={BrandType.ENLARGED} />);
+  .add('Footer', () => <Footer />);